Step R22: adopt the four ComplexNavigations bases on Tier B (bare)
Move ComplexNavigationsQuery, ComplexNavigationsCollectionsQuery and their
two shared-type siblings from ADR-009 Tier A (InMemory) to Tier B (SQLite),
under the Group C policy. All four relational bases are 11 to 19 lines and
add no test methods of their own: the two non-collection ones override a
single client-evaluation test and swap in RelationalQueryAsserter, and the
collections ones add nothing and one join-key message respectively.

Neither the core bases nor the relational ones declare UseTransaction or
call ExecuteWithStrategyInTransactionAsync, checked at both levels of the
chain rather than assumed, so no transaction override is needed.

The move deletes the fourteen InMemory-limitation overrides the two Tier A
collections classes carried for non-composed GroupBy, and the two
Correlated_projection_with_first stubs that returned Task.CompletedTask.
It also clears two entries from the compliance test's ITestSqlLoggerFactory
assertion, because both relational fixture bases implement it.

Adopted bare, per CLAUDE.md's measure-first rule: the SQLite override
subset follows in the next commit, once the run says which tests are red
and for what reason.
